Remove applyEditorModelChanges, add EditorModel.applyChanges
